Move some service system APIs to `private` module
Make it easier to control their visibility through feature flags.
